+++ src/SpeakEasy.IntegrationTests/Controllers/ReservationsController.cs
@@ -15,5 +15,15 @@ namespace SpeakEasy.IntegrationTests.Controllers
 
             return new HttpResponseMessage(HttpStatusCode.NotFound);
         }
+
+        public HttpResponseMessage Put(int productId, int priceIncrease)
+        {
+            if (priceIncrease > 100)
+            {
+                return new HttpResponseMessage(HttpStatusCode.Created);
+            }
+
+            return new HttpResponseMessage(HttpStatusCode.BadRequest);
+        }
     }
 }