+++ src/SpeakEasy.Specifications/PostRequestSpecification.cs
@@ -1,5 +1,4 @@
-﻿using System.Net;
-using Machine.Fakes;
+﻿using Machine.Fakes;
 using Machine.Specifications;
 
 namespace SpeakEasy.Specifications
@@ -14,6 +13,40 @@ namespace SpeakEasy.Specifications
 
             It should_have_null_body = () =>
                 request.Body.ShouldBeOfType<PostRequestBody>();
+
+            static PostRequest request;
+        }
+
+        [Subject(typeof(PostRequest))]
+        public class when_building_request_url_with_object_body
+        {
+            Establish context = () =>
+            {
+                var resource = new Resource("http://example.com/companies");
+                resource.AddParameter("makemoney", "allday");
+
+                request = new PostRequest(resource, new ObjectRequestBody(new { }));
+            };
+
+            It should_generate_query_params = () =>
+                request.BuildRequestUrl(new CommaSeparatedArrayFormatter()).ShouldEqual("http://example.com/companies?makemoney=allday");
+
+            static PostRequest request;
+        }
+
+        [Subject(typeof(PostRequest))]
+        public class when_building_request_url_with_post_request_body
+        {
+            Establish context = () =>
+            {
+                var resource = new Resource("http://example.com/companies");
+                resource.AddParameter("makemoney", "allday");
+
+                request = new PostRequest(resource, new PostRequestBody(new Resource("foo")));
+            };
+
+            It should_not_generate_query_params = () =>
+                request.BuildRequestUrl(new CommaSeparatedArrayFormatter()).ShouldEqual("http://example.com/companies");
 
             static PostRequest request;
         }